wip: moved status util to global util
