+++ src/screens/validator_details/components/validator_overview/utils.ts
@@ -1,28 +1,6 @@
-export const getStatusTheme = (status: number, jailed: boolean) => {
-  const results = {
-    status: 'na',
-    theme: 'zero',
-  };
+import { getValidatorStatus } from '@utils/get_validator_status';
 
-  if (status === 3) {
-    results.status = 'active';
-    results.theme = 'one';
-  } else if (status === 2 && jailed) {
-    results.status = 'jailed';
-    results.theme = 'two';
-  } else if (status === 2 && !jailed) {
-    results.status = 'unbonding';
-    results.theme = 'three';
-  } else if (status === 1) {
-    results.status = 'unbonded';
-    results.theme = 'zero';
-  } else {
-    results.status = 'unknown';
-    results.theme = 'zero';
-  }
-
-  return results;
-};
+export const getStatusTheme = getValidatorStatus;
 
 export const getCondition = (condition: number, status: number) => {
   let result = 'na';
